feat: added support for integration delays.
Integration delays can happen due to how Maybenot is integrated, e.g.,
with Maybenot in user space and WireGuard in kernel space.

There are three types of integration delays:

- The *reporting* delay is the time between an event being created by
the integrated protocol and the event being reported (trigger_events) to
Maybenot. For example, this could be the time it takes to go from kernel
space to user space.

- The *trigger* delay is the time it takes for the integration to
perform a scheduled action. For example, suppose an action is scheduled
for time T. In that case, the trigger delay is added to T. This is
important for capturing async integrations, where a zero timeout on an
action to send padding would still take some (tiny) time to execute.

- The *action* delay is the time between the integration taking action
and the action happening. For example, if a padding packet is to be
sent, user space might need to signal to kernel space to craft one.

The delays are expressed as a sequence of bins (quantile binning /
equal-frequency binning), where each bin has a probability of being
selected. This format was motivated by extracting complex distributions
from arbitrary Maybenot integrations and then creating an efficient way
to sample those distributions. Efficient sampling is extra important
because it's in the simulator's hot path.

As part of this work, I started refactoring the network abstraction to
clean up the API.